@@ -1,10 +1,19 @@
 [package]
 name = "shfl"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 license = "MIT"
+
+exclude = [
+    ".github",
+]
 
 [dependencies]
 anyhow = "1.0.98"
 clap = { version = "4.5.40", features = ["derive"] }
 ratatui = "0.29.0"
+
+[profile.release]
+codegen-units = 1
+lto = "fat"
+strip = "symbols"
